fix -Wshorten-64-to-32 warning
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@38171 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/mididrv/midi-unix-drv.c
@@ -405,7 +405,7 @@ static int mididrv_alsa_in(uint8_t *b)
 
     /* Decode Alsa event into raw bytes */
     snd_midi_event_reset_decode(midi_event_parser);
-    alsa_err = snd_midi_event_decode(midi_event_parser, buf, sizeof(buf), ev);
+    alsa_err = (int)snd_midi_event_decode(midi_event_parser, buf, sizeof(buf), ev);
     if (alsa_err < 0) {
         log_error(mididrv_log, "could not decode midi event: %s\n", snd_strerror(alsa_err));
         return -1;
